contact.md: update form URL
Formspree has recently deprecated "legacy forms", i.e. forms where the endpoint URL contains the target email address. This was a clever hack so that users would not have to create a Formspree account in order to create and use forms, but it has the downside that it exposes the user's email address to spammers.  Legacy forms also rely on an accurate HTTP referrer header, which browsers (notably Chrome) may no longer be providing for security/privacy reasons. For further details, see:  https://help.formspree.io/hc/en-us/articles/360056076314

To address this issue, I had to create a Formspree account and create a new form.